requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.0
flask==3.0.0
werkzeug==3.0.0
gunicorn==21.2.0

# AI & LLM
# ollama==0.1.0  # Optional - Install separately from https://ollama.ai

# Text-to-Speech
# piper-tts==1.2.0  # Optional - Install separately if needed

# Video Processing (Optional - install if needed)
# ffmpeg-python==0.2.0
# pillow==10.1.0
# moviepy==1.0.3
# opencv-python==4.8.0.76

# API Clients (Optional - install if needed)
google-auth==2.25.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
# tweepy==4.14.0
# praw==7.7.0

# Image Processing (Optional - install if needed)
# replicate==0.18.0

# Utilities (Optional - install if needed)
# pydantic==2.5.0
# colorama==0.4.6
# tqdm==4.66.1
# PyYAML==6.0.1

# Data Processing (Optional - install if needed)
# pandas==2.1.3
# numpy==1.26.2

# Async & Concurrency
aiofiles==23.2.1
boto3==1.30.0
psycopg2-binary==2.9.7
google-cloud-storage==2.13.0
google-cloud-firestore==2.12.0
